% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_grad_laplace.R
\name{nll_grad_laplace}
\alias{nll_grad_laplace}
\title{Function to compute Laplace-Approximation and gradient}
\usage{
nll_grad_laplace(
  parm,
  mod_list,
  yis,
  Zis,
  Wis,
  Xis,
  X,
  W,
  n_clusters,
  clust_ind,
  parm_table,
  latent_error_dist
)
}
\description{
Function to compute Laplace-Approximation and gradient
}
